requirements.txt
pyautogui
pytesseract
pillow
numpy
opencv-python
termcolor